tri1132: A user needs to be able to select a 2009 (or any other future date thru 2099) date for any date field in TriSano
"How much more work would it be to have it programmed into the app to updated automatically each year?"
-Basically, it was already implemented like this and the reason 2009 didn't exist is because it's not 2009 yet.  It's letting you select a date equal to or less than the current year; that is to say, when 2009 rolls around, it's going to have 2009.

However, pregnancy was a bit funny because it still only let you select up to the current year.  So in december, you have less than a month to be pregnant.  Now it gives you the current year +1.



git-svn-id: https://trisano.csinitiative.net/svn/trisano/top-level/trunk@1955 b21a84b9-e453-0410-92a3-c4f515f7ae45