Support arbitrary measure aggregators in Mondrian
This commit permits Mondrian (with some sleight-of-hand) to support as measure
aggregators any aggregates PostgreSQL supports. We need this so we can
calculate standard deviations, medians, modes, etc. The mechanics involved is
tricky, because Mondrian doesn't provide an easy mechanism to do what should be
an easy task. More extensive comments are available in warehouse_init.sql, and
an example measure lives in TriSano.OLAP.xml.

To make this work, existing installations will need to re-run
warehouse_init.sql, after changing references to "trisano_ro" to the warehouse
PostgreSQL user.