Began update of CSV export spec for assessments
Tried adding appropriate test coverage of assessment CSV export but
found the test suite to be poorly designed.  Sadly, it re-implements the
logic it is trying to test.  I've left the test as pending, but left the
work I started in place as it does not affect the other tests.

This relates to PT ticket 28636187, A user can export an Assessment
Event to CSV